Move code from Bloodmasters to CodeImp.Bloodmasters
