+++ Source/DedicatedServer/Bloodmasters.DedicatedServer.csproj
@@ -4,8 +4,8 @@
         <OutputType>WinExe</OutputType>
         <TargetFramework>net7.0</TargetFramework>
         <PublishDir>$(ServerPublishDir)</PublishDir>
+        <RootNamespace>CodeImp.Bloodmasters.DedicatedServer</RootNamespace>
     </PropertyGroup>
-
 
     <Target Name="CopyConfigFile" BeforeTargets="Build"
             Inputs="..\Config\bmserver.cfg"
